requirements.txt
Django==3.2
djangorestframework==3.13.1
psycopg2-binary==2.9.2
gunicorn
dj-database-url
whitenoise